♻️ refactor: dissolve galax.utils into the portions that use it
`galax.utils` was the shared grab-bag that would have forced a `galax-core`
distribution to exist. Every module in it had a single consuming portion, so
each moves to that portion and the package is deleted outright. No re-export
shims: `galax.utils` is gone.

  - `dataclasses`     -> `galax.potential.dataclasses`  (public, sole consumer)
  - `loop_strategies` -> `galax.dynamics.loop_strategies` (public, sole consumer)
  - `_jax`            -> `galax.potential._src.jax`
  - `_boundinstance`  -> `galax._boundinstance` (private portion, to be spun off)
  - `_shape`          -> split, see below

`_shape` served three portions, so it splits by function rather than moving
whole: `batched_shape` / `vector_batched_shape` to `galax.coordinates._src.shape`
(the lowest portion using them) and `expand_arr_dims` / `expand_batch_dims` to
`galax.potential._src.shape` (its only consumer).

Drop `_jax.quaxify` entirely. It was a three-line wrapper adding only a `cast`
over `quax.quaxify`, which this codebase already imports directly in
`coordinates/_src/{psps,pscs}/register_vectorapi.py`.

Tests follow their subjects: `tests/unit/utils/test_shape.py` splits across
coordinates and potential, `test_jax.py` moves to potential, and the
`galax.utils.__all__` smoke test is deleted.
