+++ src/galax/dynamics/_src/experimental/stream.py
@@ -20,8 +20,8 @@ from dataclassish.converters import dataclass
 
 import galax.dynamics._src.custom_types as gdt
 import galax.dynamics.custom_types as gt
+import galax.dynamics.loop_strategies as lstrat
 import galax.potential as gp
-import galax.utils.loop_strategies as lstrat
 from .df import AbstractKinematicDF, Fardal2015DF
 from .integrate import integrate_orbit
