+++ src/galax/dynamics/_src/orbit/solver.py
@@ -29,10 +29,10 @@ import galax.coordinates as gc
 import galax.dynamics._src.custom_types as gdt
 import galax.dynamics.custom_types as gt
 from .field_base import AbstractOrbitField
+from galax.dynamics import loop_strategies as lstrat
 from galax.dynamics._src.solver import AbstractSolver, SolveState, Terms
 from galax.dynamics._src.utils import parse_saveat, parse_to_t_y
 from galax.dynamics.fields import AbstractField
-from galax.utils import loop_strategies as lstrat
 
 BBtQParr: TypeAlias = tuple[gdt.BBtQarr, gdt.BBtParr]
