--- src/galax/utils/__init__.py
@@ -1,5 +1,0 @@
-"""galax: Galactic Dynamix in Jax."""
-
-__all__ = ["dataclasses", "loop_strategies"]
-
-from . import dataclasses, loop_strategies
