--- tests/smoke/utils/test_package.py
@@ -1,6 +1,0 @@
-from galax import utils
-
-
-def test__all__() -> None:
-    """Test that `galax.utils` has the expected `__all__`."""
-    assert utils.__all__ == ["dataclasses", "loop_strategies"]
